Remove dimensions as a placeholder
Right now we only can inform dimensions into placeholders.
That means that we need to add all the dimensions name/values
into chunking keys for the buffer.

I'm removing it, so we can inform the dimensions as a record value.
We will send them to Cloudwatch directly with minimum transformation.